seperate the utils file
