+++ src/utils/win32-beep.psl
@@ -1,0 +1,3 @@
+[console]::beep(300,400)
+[console]::beep(300,400)
+[console]::beep(300,900)
